Adding communication between LoginScene and SelectCharacterScene
