+++ front/src/Phaser/Player/Animation.ts
@@ -1,14 +1,5 @@
 import {Textures} from "../Game/GameScene";
 
-interface AnimationData {
-    key: string;
-    frameRate: number;
-    repeat: number;
-    frameModel: string; //todo use an enum
-    frameStart: number;
-    frameEnd: number;
-}
-
 export enum PlayerAnimationNames {
     WalkDown = 'down',
     WalkLeft = 'left',
@@ -16,34 +7,4 @@ export enum PlayerAnimationNames {
     WalkRight = 'right',
 }
 
-export const getPlayerAnimations = (name: string = Textures.Player): AnimationData[] => {
-    return [{
-        key: `${name}-${PlayerAnimationNames.WalkDown}`,
-        frameModel: name,
-        frameStart: 0,
-        frameEnd: 2,
-        frameRate: 10,
-        repeat: -1
-    }, {
-        key: `${name}-${PlayerAnimationNames.WalkLeft}`,
-        frameModel: name,
-        frameStart: 3,
-        frameEnd: 5,
-        frameRate: 10,
-        repeat: -1
-    }, {
-        key: `${name}-${PlayerAnimationNames.WalkRight}`,
-        frameModel: name,
-        frameStart: 6,
-        frameEnd: 8,
-        frameRate: 10,
-        repeat: -1
-    }, {
-        key: `${name}-${PlayerAnimationNames.WalkUp}`,
-        frameModel: name,
-        frameStart: 9,
-        frameEnd: 11,
-        frameRate: 10,
-        repeat: -1
-    }];
-};
+
